Allow users to override feature detection (V2)
Changes since V1:
 - Changed the environment variable name and format for specifying overrides.

In some cases (such as a distro kernel with backported features) libhugetlbfs
will be wrong about the presence of a feature.  In these rare cases it will be
useful to tell the library whether or not the feature is present in the running
kernel.

All feature overrides will be specified in the HUGETLB_FEATURES environment
variable.  The required format is: "[no_]feature1[,[no_]feature2] ... ".  The
"no_" prefix will mark a feature as absent.  Otherwise the named feature is
marked present.

Signed-off-by: Adam Litke <agl@us.ibm.com>