Recover gracefully when ELF remapping fails due to insufficient hugepages
When remapping text or data, segments are prepared in a child process and
later remapped. The status code from the segment preparation is not checked
and in the event of failure to copy text or data, the main process gets killed
later when it remaps a useless segment with no data. This patch checks the
return values from prepare_segment() and recovers gracefully in the event
prepare_segment() fails. The impact for tests is that remapping tests that
previously received a SIGKILL or SIGBUS when the hugepage pool was too small
will instead FAIL with the message "small_bss is not hugepage" or similar.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>