Add new environment variable to specify the default hpage size to use
When multiple huge page sizes are available, libhugetlbfs behavior is to use
the huge page size reported in /proc/meminfo as the default size.  Allow other
sizes to be used as the default via a new environment variable:
HUGETLB_DEFAULT_PAGE_SIZE.  If this variable is specified, libhugetlbfs will
attempt to find a mounted filesystem of this size and use it as the default.
If this is unsuccessful, the first size that was found is used.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>