Do not allow zero as an explicit page size
As originally written, the functions hugetlbfs_find_path_for_size() and
hugetlbfs_unlinked_fd_for_size() accepted 0 as a valid page size.  A page size
of zero was interpreted to mean the default size (as defined by libhugetlbfs,
not the kernel).

This behavior is undesirable in pagesize-specific functions for two reasons.
First, the libhugetlbfs default size can be nebulous since it depends on what
page sizes are available, which sizes have appropriate mount points, and the
value of an environment variable.  This amount of variability adds uncertainty
to what should be an explicit API.  Second, the functions hugetlbfs_find_path()
and hugetlbfs_unlinked_fd() already exist as a means to delegate page size
selection to the library.  It is my view that all functions that take a page
size paramater shall expect a valid page size.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andy Whitcroft <apw@shadowen.org>