libhugetlbfs: Print summary of testsuite results
O> On Wed, Sep 24, 2008 at 12:00:31AM +0100, Eric B Munson wrote:
> > On Wed, 20 Aug 2008, David Gibson wrote:
> >
> > > This patch adds some tracking to the run_tests script which counts the
> > > total number of tests, and the number of passes, failures and other
> > > results.  This is printed as a summary once the script is finished, so
> > > you can see at a glance if the tests ran clean or not.
> > >
> > > Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
> > >
> >
> > David,
> >
> > Do you have the time to clean up the whitespace and repost?  If not I
> > can give it a shot and apply this patch.
>
> Uh.. sorry, I've lost track of this.. what was the problem with the
> whitespace?

Sorry, re-examined the thread and figured it out.  This should fix the
tab problem:

libhugetlbfs: Print summary of testsuite results

This patch adds some tracking to the run_tests script which counts the
total number of tests, and the number of passes, failures and other
results.  This is printed as a summary once the script is finished, so
you can see at a glance if the tests ran clean or not.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>