Specify a page size with HUGETLB_MORECORE (V2)
For systems with multiple huge page sizes, it is desirable to select which page
size should be used for backing the heap.  This patch overrides the
HUGETLB_MORECORE environment variable so that it may be set to a desired huge
page size.  Once applied, the following values are recognized in
HUGETLB_MORECORE:

	"no"	- Disables the feature entirely
	"y[es]"	- Enable the feature and use the default huge page size
	<size>	- Use the specified huge page size

The size method can be specified in bytes or in kilobytes, megabytes, or
gigabytes by appending K, M, or G respectively.  It is an error to specify an
invalid, unsupported, or otherwise unconfigured huge page size.

Changes since V1:
 - Fixed the CodingStyle of a comment
 - Make HUGETLB_MORECORE environment parsing disregard string case.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>