shmoverride test fix
The shmoverride tests currently hang when run as a user.  This is because the
cleanup routine tries to reset the nr_hugepages value to what is was at the
beginning of the test.  This will fail with bad CONFIG which will try and
call cleanup again.  This patch initialises saved_nr_hugepages to -1 and
does not restore the value in cleanup if it is <= 0.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>