shm: Shared memory always uses the meminfo page size
This patch fixes two intertwined issues:

When the SHM_HUGETLB flag is passed to shmget(), the system default huge page
size is used to back the shared memory segment.  Unlike with mmap() there is
not yet a way to use an alternate huge page size for shared memory segments.
This exception must be taken into account by the shmget override function -- it
cannot call gethugepagesize() but must discover the effective huge page size by
direct examination of /proc/meminfo.  Make the shmoverride test case discover
the huge page size in the same manner.

The shmoverride_* test case does not always link against libhugetlbfs.
Therefore, it cannot use the new mechanism for manipulating pool counters.
Unfortunately, this means that the logic to read /proc/meminfo must be
duplicated in order for shmoverride_unlinked to build as a NOLIB test.

[apw@shadowen.org: consolidate makefile updates in one place]
Signed-off-by: Adam Litke <agl@us.ibm.com>