hugeutils: size_to_smaller_unit should allow 2GB pages
When page sizes become very large they may overflow a long as we
convert them from giga-, mega-, or kilo-bytes to bytes.  When doing the
conversion if we hit an error we return LONG_MAX.  On 32bit this is 2GB
is is potentially a sensible page size for a 32bit application.  As page
size is a long we have all of the negative range available so it seems
logical to use a negative value to indicate error.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>