utils: Make pool counter manipulation code part of the library API
The need to set and get pool counters for multiple huge page sizes is about to
extend beyond just the libhugetlbfs tests.  Move the counter code into
hugeutils.c and make the top-level functions part of the libhugetlbfs pool and
mount configuration API.  This involves altering some function names and
prologues to make them more consistent with the existing libhugetlbfs API.

We create a new libtestutils.c for test case utilities that depend on linkage
to libhugetlbfs.

NOTE: This patch will cause the shmoverride test to fail compilation.  This is
fixed up by the next patch.

[apw@shadowen.org: consolidate makefile updates in one place]
Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andy Whitcroft <apw@shadowen.org>