hugeadm: allow pools to be resized
Allow resize of specific pools via the --pool-minimum-adjust and
--pool-maximum-adjust commands.  The first affects the number of static
pages in the pool, the second affects the number of overcommit pages
for the pool.  These are exposed as lower and upper bound on the pool
to match the display in --pool-list.  Adjustments are either absolute,
or relative with a + or - prefix.  For example the following incantion
increases the 2MB page pool minimum by 10 pages.

	--pool-minimum-adjust 2048kb:+10

Note we must handle the possiblity that the pool minimum may not increase
to the value requested should memory not be available.  In this case we must
recalculate the overcommit.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>