Allow individual installation of manual pages, utilities or library with linker scripts
Currently the install target installs the library, utilities and the manual
pages. In the situation where the user has downloaded libhugetlbfs for
the utilities, it is relatively difficult to build utilities that use the
main system paths, but do not override the distribution-supported version
of libhugetlbfs.

This patch allows a user to do something like

$ make PREFIX=/usr
$ make install-bin
$ make install-man

and get the utilities and manual pages using the distribution-supported
library.

The install target still installs everything and should be functionally
equivilant. The only difference is a clarification in output that
INSTALL32 is really INSTALL-LIB32

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>