move to a new library local idiom
Currently we have three types of function:

  file local -- marked static in the normal way,
  library local -- external but prefixed with __lh_
  library exported -- external and listed in the library.lds file

While the library prefix works, it does not allow functions to trivially
move from file local to library local as all references to the function
have to be modified to the new name.

This patch introduces a new idiom.  When a function is intended to be
library local it is already necessary to declare that function in the
libhugetlbfs_internal.h, if we also add a single define for that function
adding the __lh_ prefix (as below) then all other references including the
original definition may use the original name unchanged but the function
will remain unexported:

	#define hpool_sizes __lh_hpool_sizes
	extern int hpool_sizes(struct hpage_pool *, int);

This patch converts all current library local functions to this new idiom.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>