elflink: Fork to prepare segments when sharing too
The functions find_or_prepare_shared_file() and obtain_prepared_file() both
call prepare_segment().  The former when sharing is enabled and the latter when
not.  While it was deemed beneficial to fork a child process to run
prepare_segment(), it was only implemented for the case when sharing is not
enabled.  Add a new function fork_and_prepare_segment() so that forking can be
easily done in both instances.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>