add a new private utilities library
Add a new private utilities library consisting of various useful helpers
normally hidden within libhugetlbfs.  Extend the library local marker idiom
to include a private utilities marker __pu_.  When we build libhugetlbfs
these are forced local when building libhugetlbfs_privutils these are the
only routines exported.  This makes it very hard for the two libraries to
interfere with each other when both are linked to the same binary, which is
particularly important when testing the library; we must test the real one.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>