hugeadm: cleanup pool size adjustment for kernels without overcommit
In older kernels we have no overcommit support.  This means that there
can never be a minimum/maximum spread.  Also it means we do not have
overcommit counters on read.

Basically we make --pool-pages-max error indicating that there is no
overcommit support.  When we attempt to change the minimum where there
is no overcommit support we simply always update the maximum to match.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>