privutils: expose the kernel default page size
We use the kernel default page size all over the place.  We use it
approximatly 6 times when setting the boundaries on a single pool.
Currently every place looks this value up in /proc/meminfo which is
very costly.

As this value is essentially static per boot we can cache this value.
Create a new routine kernel_default_hugepage_size() which looks this
up and caches the result.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>