tests: gethugepagesizes -- switch meminfo override to intercepting open
The current override for the contents of meminfo via a replacement
read_meminfo will not work now that these functions are in a library.
Switch to overriding the files returned from open by overriding that
instead.  This is actually a superior way to test as it also allows us
to test the library implementation of read_meminfo.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>