gethugepagesizes: fix selection of counting mode
gethugepagesizes has two modes: a counting mode which tells us exactly
how many entries we would need; and a lookup mode which fills in a
supplied buffer.

The counting mode is defined as triggering when we pass in both a zero
element count and a NULL buffer pointer.  Currently we enter this mode when
the element count is zero which causes it to trigger at the edge case where
the caller (pointlessly) passes in a buffer with an element count of zero.
Fix behaviour to match the documentation.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>