Handle the case where lib64 and lib32 resolve to the same place
On recent 64-bit Debian systems, /usr/lib64 is a symbolic link to
/usr/lib and 64-bit libraries are installed to /usr/lib/$ARCH-linux-gnu.
In this case, the 64-bit install of libhugetlbfs overwrites the 32-bit
version of the library. It's different in Ubuntu where lib can be the
64-bit directory, lib32 the 32-bit one and lib64 a symbolic link to lib.

When installing with PREFIX=/usr, this leads to all sorts of hilarity and the
bizarre situation where the 64-bit install overwrites the 32-bit install. How
each distro arranges 32/64 bit libraries in /usr varies considerably so
rather than trying to figure out what to do, inform the user when the library
installation paths are going to the same place and allow them to manually
specify the correct library path names within the prefix.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>