tests: Infrastructure for expected test failures (V2)
The libhugetlbfs test suite has various tests for which a failure (under certain circumstances) is expected and not indicative of an error.  It would be desirable to specially annotate these cases to eliminate the need for the tester to know which failures are known and which could be bugs.

This patch creates some test harness infrastructure that can be used to achieve this goal.  How does it work?  If a test case needs expected failure validation, a shell script (named <testcase>.sh) is placed in the tests/ directory.  The test invocation in run_tests.sh is changed to execute the shell script instead of the c program directly.  The shell script must perform any required discovery, execute the real test case, and interpret the result.

Comments?

Changes since V1:
 - Revamped the build so wrappers properly depend on their test cases

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>