add a new export type for main library probe points for tests
When testing the library we sometimes want to influence data which becomes
cached within the library.  To allow this we need to reset that cached data
to ensure it is re-read from our test data instead of the real system.
This change adds the concept of a test probe point to the main library.
These are external symbols but not exposed in the standard header, but
in libhugetlbfs_testprobe.h.  They are also name mangled.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>