hugetlbfs default page size is cached so expose reset for testing
When testing some gethugepagesizes we desire to fake the default huge
page size, to test both having and not having hugepage support on the
same machine.  Now that the default huge page size is cached we now
need to invalidate this cache to allow us test with various values.

Add and expose kernel_default_hugepage_size_reset() as a test probe,
and use this within the gethugepagesizes test.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>