hugectl: add support for multiple page sizes
Add support for multiple page sizes to hugectl.  Now that the library
supports specification of the backing page size to use for the various
mappings we should expose this functionality in hugectl.  This patch
adds an optional size argument to all of the segment mapping options
specifying the backing page size required.  These are converted into the
appropriate variables and passed on.  For example, the following incantation
requests text be placed in 2MB pages, data in 64KB pages and the heap in
the default size:

	hugectl --text=2M --data=64K --heap ls

Combinations which are not possible result in warnings as normal.
The sizes are not validated but passed verbatim to the library which will
then validate them.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>