elflink: Do not worry about layout for base page segments
Before doing any segment remapping, the function verify_segment_layout() does
some basic sanity checking to make sure the program segments are properly
aligned for the selected remapping mode.  For segments that will be remapped,
we must be able to align the start down to an address congruent with the target
page size and.  We must similarly align the end upward.

For segments that will not be remapped, we don't actually require such
alignment to be possible but we have been sanity checking it anyway.  This has
been okay for awhile but newer versions of binutils are producing binaries with
segments that are not aligned.  These binaries can work fine with libhugetlbfs
as long as the unaligned segments do not need to be remapped into huge pages.
Update verify_segment_layout() to skip the checks for these segments.

This issue has, so far, only been reproducible when using the B-type linker
scripts because it leaves the text and data segments untouched and creates a
new segment for a huge page BSS.  We need to ignore the boundary between the
text and data segment because it is not important.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>