Force get_huge_pages() to only deal with hugepages
It has been asserted that the name get_huge_pages() implies the API is
a close-to-kernel interface for allocating hugepages and nothing else.
In 2.1-preX, this definition does not fit as it allows fallback to base pages
for example.  This patch enforces the meaning implied by the function name by
removing the ability of the function to fallback to base pages and redefines
GHP_DEFAULT to mean "Use the default hugepage size to back the region".

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>