Add get_hugepage_region() as a malloc-like interface that uses hugepages
The get_huge_pages() API is a close-to-kernel interface for the direct
allocation of hugepages. This forces the caller to deal with alignment and
fallback to base pages where suitable. For the casual user of hugepages
that does not care for such things, this patch adds get_hugepage_region().
It allocates regions of memory that are backed by hugepages where possible
but callers are not required to align their length and can request fallback
to base pages.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>