Use wasted bytes in get_hugepage_region() to place buffers at different cache lines
In the situation a caller allocates multiple buffers with
get_hugepage_region(), all the buffers will align to the start of the
hugepage. This means the buffers are almost certainly using the same
cachelines and operations like copying between the arrays will be mysteriously
slow in comparison to buffers allocated from malloc(). In STREAM, allocating
the three arrays using get_huge_pages() was approximately 40% the throughput
of malloc().

When requested, this patch uses bytes that would normally be wasted to
offset the buffer by some random cacheline. Effort is made to select a
random cache-line to give good average performance.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>