tests: fork-cow test requires 2 huge pages
In order for the child to finish without being killed, 2 huge pages are needed.
Also, handle the case where the child is terminated by catching SIGCHLD in the
parent.  This avoids a test case hang where the parent will wait around forever
on a trigger that will never be tripped.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>