Correctly handle the freeing of unaligned pointers in free_hugepage_region API
free_hugepage_region() is currently assuming that the region was backed
by hugepages. This may not be the case when GHP_FALLBACK is specified
and it was luck on the original test machines that the addresses
happened to work out. The test case also makes assumptions on alignment.

This patch makes free_huge_pages() an internal helper function called
__free_huge_pages() which can take aligned or unaligned addresses. When
it is told the address is unaligned, it'll check the mappings to see if
it was page or hugepage aligned before freeing. free_huge_pages() and
free_hugepage_region() only differ then on whether the addresses are
aligned or not.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>