Update SECTIONS from linker scripts for binutils 2.19 compatibility
When building the linker script tests on SLES 11, the resulting executable
had an invalid size for the interp section.  This is dues to a new section
that was not present in the old script.  This patch adds this section to
all of the linker scripts.

Credit goes to Alan Modra for identifying what change needed to be made

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>