Update HOWTO for NX bit workaround V3
Presently on machines that support the NX bit, 32bit programs which
were linked with the xB linker script will segfault.  This is a known
problem but will not be fixed because the linker scripts have been
deprecated in favor of the --hugetlb-align linking method.  This
patch adds two possible work arounds to the HOWTO file for this problem.

Signed-off-by: Eric Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>