hugeadm: Correct calculation of "Current"
"Current" in the output of --pool-list is meant to be the number of
hugepages currently in use, either by applications or stored on the
kernels free. What we actually report is number used + the number
reserved which is not quite right.

For example, take the situation where
	o Min = 5, Max = 10
	o no applications running => Current = 5

That is correct. 5 pages are in use because they are in the static pool.
Now, start an application that mallocs 3 huge pages and faults them

	o Min = 5, Max = 10
	o Current = Total + Reserved
	          =     5 + 0        = 5

Still correct. Now, take an application that mallocs 3 huge pages but
does *not* fault them

	o Min = 5, Max = 10
	o Current = 5 + 2            = 7

Which is wrong. There are still only 5 pages in use. Lets say the
application malloced 9 huge pages, then Current becomes larger than
Maximum which makes no sense.

This patch changes Current to be based on the value HugePages_Total.
Review on this change is important as it's easy to have missed
something.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>