Add --create-group-mounts and --create-user-mounts options to hugeadm
The --create-group-mounts option takes a group name as an
argument and creates a subdir to /var/lib/hugetlbfs/ with
that group name.  It then creates a mount point for each
available huge page size under that directory with mode
070, mounts it with the appropriate pagesize argument and
chowns it to the specified group.

The --create-user-mounts option does the same as the
group-mounts option except it takes a username, the
mode is 700, and it is chowned to the specified user.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>