Wrap the shm-fork and shm-getraw tests for expected failures
The new test suite tries to run all tests with all available page sizes.
This will be a problem with the shm-fork and shm-getraw tests as the shmget
interface will only work with the default page, this patch wraps the tests
to make the runs with pages sizes other than the default expected failures.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>