Rename vairable and change error messages in get_huge_pages to avoid confusion
Because get_huge_pages is not used exclusively for creating a heap region it
can cause confusion to have the huge fd named heap_fd and have the error
messages say talk about a heap region.  This patch changes the variable name
and error messages to a more generic "buffer".

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>