hugeadm: Retry allocation of hugepages to static pool on failure when --hard is specified
Growing the hugepage pool with --pool-pages-min can fail due to the
fragmentation within the system but one failure is not a guarantee of
future failure, particularly if a very large amount of memory has to be
reclaimed to satisfy the allocation. This patch adds a --hard flag to
retry allocations multiple times. It initially tries to resize the pool
up to 5 times. It continues to try if progress is being made towards
the resize.

Signed-off-by: Avantika Mathur <mathur@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>