Add --force-preload option
Binaries for ppc64 are already aligned properly for 64KB backed program
segments so it should be reasonable to use 64KB huge pages for these
segments without relinking, however this requires preloading libhugetlbfs.
Currently hugectl will not allow libhugetlbfs to be preloaded if huge page
backed segments are requested.  This patch introduces the --force-preload
switch that allows the user to bypass this check and preload libhugetlbfs
any time.

This is because backing text/data assumed that linking with libhugetlbfs
is required. If an application is already linked, then the library
cannot also be preloaded. In the case of PPC64, this assumption is not
true 100% of the time but hugectl has no way of automatically detecting
that case.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>