Add --verbose to hugeadm and allow the option to be specified in any order
This patch adds the --verbose/-v options to hugeadm.  When --verbose <0-99> is
specified, the libhugetlbfs verbosity level is set to that value. The verbosity
level is increased by one each time -v is specified.  If the level is
increased to 99, this sets HUGETLB_DEBUG.
The patch also also adds support for multiple options to hugeadm to be
specified in any order.  Without this the verbose option would have to be
specified before any other options.

Signed-off-by: Avantika Mathur <mathur@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>