Check if directory is already mounted
Currently hugeadm will happily remount hugetlbfs to the same directory
if --create-*-mounts is specified more than once with the same arguments.
This is a problem because the newest mount of a directory will mask the
previous one(s) causing confusion when applications attempt tp share
segements.  This patch makes all the --create-*-mounts switches check
to see if the requested directory is already mounted and if so skip
the mount request.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>