Convert meminfo value from KB to B
Meminfo reports values in KB so when we read the amount of configured
swap space and compare to the size (in bytes) of a huge page, hugeadm
can complain about not having enough swap.  This patch converts the
value read out of meminfo from KB to B before comparing to the hugepage
size.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>