hugeadm: Enable/Disable the use ofZONE_MOVABLE for the allocation of huge pages
The kernel partitions memory if kernelcore= or movablecore= is specified on
the kernel command line creating a ZONE_MOVABLE zone. This zone contains
only pages that the kernel knows it can reclaim. Unless mlock() is used,
the hugepage pool is guaranteed to be able to grow to the size of the memory
partition but the partition is not used by default as the intention of the
partition may also be to hot-remove the memory later. Whether the zone is used
or not is determined by the value in /proc/sys/vm/hugepages_treat_as_movable.

This patch adds the option to hugeadm to enable/disable the use of
ZONE_MOVABLE for the allocation of huge pages.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>