hugeadm: Adjust opt_max_adj[] for--pool-pages-max, not opt_min_adj[]
Commit f6d20b8135e0c1ca73e8ae329be47d43be261c0b appears to have broken
--pool-pages-max processing. When --pool-pages-max is used, it adjusts
nr_hugepages instead of nr_overcommit_hugepages. This leads to the crazy
situation where min has to be set to a minimum value, the max value and then
the min value again to actually configure overcommit.

The problem is that opt_min_adj[] is being manipulated when --pool-pages-max is
specified. This patch populates the opt_max_adj[] array for --pool-pages-max,
instead of opt_min_adj[].

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>