Move check for private reservations out of morecore setup
Prefaulting should be disabled when the kernel supports private reservations.
Currently this check is part of the morecore_setup function and is only made
if HUGETLB_MORECORE is 'yes'.  This is a problem for users of
get_hugepage_region because prefaulting will slow allocation down significantly
and screw with NUMA placement policies where we want to do demand faulting.
This patch moves the check for private reservations into s separate function
that is called during hugetlb_setup.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>