Add a test that mixes permissions on shared memory segments
Due to some confustion between VM_SHARED and VM_MAYSHARE in hugetlbfs,
it is possible for hugepage reservations to get leaked. By using
VM_SHARED, hugetlbfs will treat a shmem segment mapped read-only as if
it was MAP_PRIVATE.

This patch adds a test that checks if the kernel is vunerable to this bug.
Note if this test fails, the system may no longer be usable for hugepage
testing as the system will always think it has insufficient pages.

A patch is currently being tested for this bug but no fix is merged
upstream yet.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>