Add -f option to run_test.py to force dangerous tests to run
Currently the readahead_reserve and [m|f]advise tests are all skipped
if the running kernel is earlier than 2.6.30 because they can cause
the machine to hang.  This patch adds a -f switch to run_tests.py that
allows the user to override this behavior.  If -f is specified to
run_tests.py these potentially dangerous tests will be run.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>