hugeadm: Add temporary swap on disk for duration of positive pool resize when --add-temp-swap is specified
When growing the hugepage pool with --pool-pages-min, swap space is generally
needed to allocate free memory for the resize.

For systems that do not have enough or any swap configured, the user can
specify the --add-temp-swap flag.  When this flag is specified, temporary swap
space is created and activated on disk for the duration of a positive resize of
the minimum pool size.  After the resize is completed, the swap is turned off
and swapfile removed.

Signed-off-by: Avantika Mathur <mathur@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>