Clear up compile warning from hugeadm
The return value from the system() call in add_temp_swap() was
being ignored, generating compile time warnings.  This patch adds
a check of the return value to make sure that all is well.  It also
changes the ERROR reporting in add_temp_swap to WARNING in the cases
where execution is not stopped.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>