Add perf data collection
Now that PCL has been merged to mainline, the perf userspace tool can provide
light weight access to profiling data.  This patch adds a PerfCollector to the
list of collectors available to cpupcstat and it makes PerfCollector the
preferred data source, only falling back to oprofile if perf is not available
or not supported for the architecture.

Also add a switch to cpupcstat to force data collection from oprofile.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>