Adjust tool compiler to build 64 bit tools if available
The make process currently builds all lib huge tools as 32 bit binaries
regardless of the build environment.  This causes a problem on ppc64 when
attempting to deal with 16GB pages as their size is larger than a 32 bit long
long.  This patch updates the tools build to use the 64 bit compiler whenever
it is available.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>