Scale sampling of cycles and events separately
oprofile_start.sh has the option of scaling the sampling of PMU events
by some factor. The reasoning is that oprofile introduces high overhead
when sampling at it's maximum rate - somewhere between 8% and 14%
depending on the workload. However, the minimum sampling rate for CPU
cycles and events such as DTLB misses are significantly different. CPU
cycles may know need to be scaled by a factor of 4 and DTLB misses by a
factor of 16 to get acceptable performance overhead of profiling.

This patch allows CPU cycles and events to be scaled by separate values.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>