hugeadm: Report if min_free_kbytes is too small when --explain is specified
To maximise efficiency of anti-fragmentation, pageblocks must be kept free
from mixing page types. The minimum value for this is one pageblock per
zone in the system although even this might be too minimal on x86-32 where
the size of zones vary wildly.

The easiest way to keep pageblocks free from mixing is to increase the value
of min_free_kbytes. This patch compares the current value of min_free_kbytes
with a recommended value and warns if the current value is too low.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>