hugeadm: Provide an option to set a recommended value for min_free_kbytes
hugeadm --explain will detect if the value of min_free_kbytes is too small to
avoid excessive fragmentation in the system. This patch provides an option to
automatically set min_free_kbytes to a recommended value. In this version,
the recommended value is one pageblock to be kept free per zone in the
system. This may turn out to be too conservative, particularly where
there are large variances between zone sizes but it is a reasonable
starting point.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>