Change DTLB miss events in PerfCollector
The perf tool does not require read and write DTLB misses be specified
separately.  Because we care about total DTLB misses rather than if
the miss was caused by a read or a write the events should be condensed
into a single event entry.  This patch tells the perf tool to watch for
and aggregate all DTLB miss events.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>